internal/cvelist: add walkRepo skeleton
A skeleton of walkRepo is added, which looks at all of the files in the
CVE list repo and prints the CVE ID if the file has not been triaged before.

In the future, logic will be added to check if it is a CVE that needs to
be added to the vulndb or we can ignore.

Change-Id: I62d627809c7c38120b87dccc38fb00688f8ae1da
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/356392
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Roland Shoemaker <roland@golang.org>