internal/worker: add tracing
Add support for OpenTelemetry tracing of certain
important functions, using the event package.

We don't have a burning need for traces, but they
are nice to have, and this was an important exercise
for validating the approach of the event package.

Change-Id: I37d1f56f06f425f3b1eb885877a0d2f5ac85a098
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/380440
Trust: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Jonathan Amsterdam <jba@google.com>
Reviewed-by: Julie Qiu <julie@golang.org>