deploy: add README.md and kokoro.sh
A README is added to the deploy/ directory to explain what the files are
in that directory.

The contents of the internal kokoro/vulndb/dbdiff/buid.sh file is copied
to deploy/kokoro.sh to make it easier to make modifications to the
kokoro config.

Change-Id: I62e829aa881729837e7adb56eaacc4f7e6618713
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/382194
Trust: Julie Qiu <julie@golang.org>
Run-TryBot: Julie Qiu <julie@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Jonathan Amsterdam <jba@google.com>