x/vulndb: add reports/GO-2022-0391.yaml for GHSA-76wf-9vgp-pj7w
The security advisory states that the relevant header was removed
in v1.34.0 of the AWS SDK, but I needed to go back farther than
that to find a release that sets it. Leaving the fixed version
as v1.34.0, but I suspect the real version is earlier than that.

Fixes golang/vulndb#0391

Change-Id: I14c15d33b77a7f7192755939861ee648505f1ef9
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/414822
Run-TryBot: Damien Neil <dneil@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Tatiana Bradley <tatiana@golang.org>