x/vulndb: add reports/GO-2022-0274.yaml for CVE-2021-43784
The linked CVE and GHSA describe an integer overflow error
which is fixed in 1.0.3, but the Project Zero report also
describes an issue with NUL bytes in inputs that is not
fixed until 1.1.0. I've used the later version as the fix.

Fixes golang/vulndb#0274

Change-Id: I36cfc402db1e0d48da068f93c9d618cb88bbe22a
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/416335
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Tatiana Bradley <tatiana@golang.org>
Run-TryBot: Damien Neil <dneil@google.com>