internal/report: add links for CVEMetadata
Links are currently not generated for reports where the CVE is set in
the CVEMetadata section. These are now added, but using cve.mitre.org
links, because the report may be in a RESERVED state and not have a link
on NIST.

In the future, we should check for the report state to determine which
link to include.

For golang/go#54488

Change-Id: I1bd00c9ec42f3ce35fbe9950e36058606853538a
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/424001
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Julie Qiu <julieqiu@google.com>
Run-TryBot: Julie Qiu <julieqiu@google.com>
Reviewed-by: Damien Neil <dneil@google.com>