internal/database: set OSV modification times from data/version.md
When we change the generator, we want to update the modified field
of older reports to the time of the generator change.

Track generator changes in data/version.md.

When generating OSV records, use the maximum modification time of
the report YAML and data/version.md as the time of last modification.

Also change the generator to look at HEAD rather than origin/master
for modification times; it's surprising for the generator to consider
the contents of files at HEAD (in the current working directory)
but the history on a different branch.

Change-Id: Ibc9e78667910d8013a5d05e2273adb9c48ac6269
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/424998
Reviewed-by: Julie Qiu <julieqiu@google.com>
Run-TryBot: Damien Neil <dneil@google.com>