all: refactor links to more closely match OSV
Change the 'links' section of reports to be a list of links,
each with an associated type. The link types are the same as
in OSV ("pr" and "commit" are merged into a single "fix" type,
and we now have a new "report" type for issue tracker links).
It is now possible to link to multiple PRs/commits.

Change the report parser to convert from the old report links
section to the new references section.

A followup CL will rewrite the data/ directory into the new style.

Change-Id: Ia0df42c817da103239289a93c4b52081238a35c3
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/426035
Run-TryBot: Damien Neil <dneil@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Tatiana Bradley <tatiana@golang.org>