cmd/vulnreport: clear derived symbols before rederiving them
We were using the previous value of the derived_symbols as
part of the input symbol set when computing derived symbols.
Once a derived symbol was added to the set, it would remain
forever even if the root symbol it was derived from was
removed.

Change-Id: I7fa96a7e1e5b1a1505ee8bf7eb22cf9df8102831
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/427336
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Julie Qiu <julieqiu@google.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
Run-TryBot: Damien Neil <dneil@google.com>