internal/worker: fix readAfterWrite bug
When triaging GHSAs, make a list of all GHSAs that need to
be created or updated and do the database writes all at once.

We need to do this to avoid a readAfterWrite bug in which
we attempt to read CVE records after writing GHSA records.

Change-Id: Ib5745e19f3eb9460178d02f7a7e3bc6aa024a848
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/432995
Run-TryBot: Tatiana Bradley <tatiana@golang.org>
Reviewed-by: Damien Neil <dneil@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Tatiana Bradley <tatiana@golang.org>