all_test: add test to compare data/osv and vuln.go.dev
The test is currently skipped because it fails on TryBots.

We plan to add it as a step in the deployment process to make sure that
not reports are ever deleted.

For golang/go#56139

Change-Id: If481a607174efceb73d22d6438d0465f035d40c0
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/440635
Reviewed-by: Julie Qiu <julieqiu@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Auto-Submit: Julie Qiu <julieqiu@google.com>
Reviewed-by: Tatiana Bradley <tatiana@golang.org>
Run-TryBot: Julie Qiu <julieqiu@google.com>