internal/report: properly mark defaultStatus and link to go advisory when saving CVE 5.0 JSON records
Previously, if a record did not have any versions associated with it, the default status would be marked as "unaffected" when it should be "affected". This change reflects that.
Additionally, a link to the record's go advisory is now stored in the references section of the CVE 5.0 JSON.

Change-Id: I55be77062638d34ba273b0c32799639f982cc266
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/444876
Run-TryBot: Maceo Thompson <maceothompson@google.com>
Run-TryBot: Tatiana Bradley <tatiana@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Tatiana Bradley <tatiana@golang.org>