cmd/vulnreport, internal/cvelistrepo: refactor FetchCVE to make repeated calls more efficient
We previously passed in a path to a repo which forced the vulnreport command to clone/open the github repo every time we added a new report. Now, we use a git.Repository pointer instead of path to git repo (as a string) as an arg for FetchCVE and in vulnreport/main.go's createCfg struct. When adding multiple reports in a single command, that clone/open logic will only occur once.

Change-Id: I92dd6dbf6fe52d704f27663dca98a5668ec8cf56
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/445636
Reviewed-by: Tatiana Bradley <tatiana@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Maceo Thompson <maceothompson@google.com>