cmd/vulnreport, internal/reports: fix xref bug in which no xrefs were found
Fixes a bug in vulnreport xref in which no cross-references were found due to incorrect parsing of filepaths. Filepaths are now parsed in a function in report.go and errors are returned if a filepath cannot be parsed.

Change-Id: I862b896de1ae89c49033aa98cc2f17191831ab72
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/447158
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Tatiana Bradley <tatiana@golang.org>
Reviewed-by: Damien Neil <dneil@google.com>
Run-TryBot: Tatiana Bradley <tatiana@golang.org>
Reviewed-by: Tim King <taking@google.com>