internal/database: update and test Load
Add a new struct, Database, which is an in-memory representation of
a Go vulnerability database. Update Load to return this struct.

Also adds a test database folder which is used by tests for Load, and
can be used in tests for Generate.

For golang/go#56417

Change-Id: I62d882722186ee96846b646cf88d4320a4f3a307
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/448841
Reviewed-by: Tatiana Bradley <tatiana@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Tatiana Bradley <tatiana@golang.org>
Reviewed-by: Damien Neil <dneil@google.com>