cmd/vulnreport: move find GHSA logic to newReport
This allows us to re-use the addGHSA function. Note this does slightly
change the behavior of vulnreport create, as the added GHSAs are not
taken into account when initially creating the report. This does not
matter much with the current implementation, as we arbitrarily choose
one alias to create the report based on.

Change-Id: Ia99eac8aaec603f5fd44f7b9d017957f8147fe06
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/467295
Reviewed-by: Tim King <taking@google.com>
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: David Chase <drchase@google.com>