internal/database: add full validation for old db in Validate
Change RawLoad to Load in v1 Validate so that the existing database
is validated against the v1 spec before deploy. This was previously
"RawLoad" because on first deploy the existing database was not a valid
v1 database.

For golang/go#58816

Change-Id: I45728266ae0b59b6fd486d196a2a91a712aaf12e
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/473195
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Damien Neil <dneil@google.com>