cmd/vulnreport, internal/report: add support for summary field in YAML
Adds a field, "summary", which corresponds to the OSV "summary" and
CVE "title" field. This field is pulled automatically from GHSAs in
"vulnreport create".

Currently, this field is not required and is not
populated in the OSV/CVE conversion. Introducing it now will make
it easier for us to begin publishing this field later, to reduce
the backfill burden.

For golang/go#56443

Change-Id: Ib93efad656daeac4b13a97d83d46952dbced14b5
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/475336
Reviewed-by: Julie Qiu <julieqiu@google.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>