internal/database: fix bug in latestFixedVersion
Fixes a bug in which latest fixed version was computed incorrectly
when a vuln was re-introduced and never subsequently fixed. (This is
an uncommon case, and never appeared in our actual database).

Also adds a test for this case.

Change-Id: I7a88f16a4fef079d8bb23f6a2e9ffd7dd57db30a
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/487455
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Damien Neil <dneil@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>