cmd/vulnreport: replace dependency on vulncheck with its copy
vulncheck is not exported any more so we cut ties with it here. This CL
replaces calls to vulncheck with a slightly modified copy of parts of
vulncheck.

The copy contains just enough logic to compute derived symbols, i.e.,
entries of call graph leading to vulnerable symbols. Hence, the copy
manifests as an internal package of vulnreport called vulnentries.
Other logic of vulncheck is not copied.

The copied code of vulncheck is unlikely to change in the future and has
not change pretty much from its outset in vuln. This will hopefully
minimize the maintenance burden of vulnreport going forward.

vulnreport now does not depend on golang.org/x/vuln.

Change-Id: Ie5e7bab639ad69144a9be64f1899c722c13db37c
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/496475
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>