internal/worker: use function Aliases instead of x/vuln client
To remove the dependency on the now unsupported x/vuln client, clone
the vulndb repo and call Aliases to find all aliases. This is actually
an improvement, as the worker will now take into account excluded aliases
when determining if something is a duplicate.

With this change, we have now completely removed the dependency of this
repo on x/vuln.

Fixes golang/go#60116

Change-Id: I12a837745d4eb2cc62cdb44522a52e2d016b4b6c
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/497039
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Damien Neil <dneil@google.com>
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>