internal/database: validate current database before deploy
Validate current database before deploy so we don't accidentally
deploy on top of an invalid database. (If this fails, a human should
step in and manually fix it).

This is done by calling the function "Load" (which loads and validates
a database) instead of the "RawLoad" function (which loads a database
without validating it).

Change-Id: If50b54d95a8f439f1eb51dfb1d1cf3690a08b5a4
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/506562
Reviewed-by: Damien Neil <dneil@google.com>
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>