internal/genericosv: add function to convert OSV to Go Report
Add function ToReport which converts a generic OSV entry to a Report
(the representation for our YAML reports). To start, this function
simply translates between the formats and doesn't do anything clever.

This change also adds a number of real GHSAs that can be used as test
cases, and to see how the function behaves on real data as it evolves.

Lint errors are added as notes to generated reports, so that we can
more easily target areas for improvement.

For golang/go#61769

Change-Id: Ifd99796f96aa662e887a643276b3b2d7456e826b
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/515399
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Damien Neil <dneil@google.com>
Run-TryBot: Tatiana Bradley <tatianabradley@google.com>