internal/proxy, all: refactor proxy tests
Move code for proxy testing to the proxy package and re-use it across
the module. This allows us to remove many real calls to the proxy in tests,
which are brittle and require a network connection.

Additionally, refactor testing in genericosv.ToReport so that each
sub-test has it's own stored proxy responses JSON blob.
This will make it easier to see which tests are affected
by a change, and to isolate proxy responses between tests.

Change-Id: Icf52cff82ded3bb4627d55448051d230310eb670
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/524076
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>