internal/{report, genericosv}: add unsupported_versions field to YAML
Add a field, unsupported_versions to YAML reports, which holds version
types that we don't support in our reports. This field serves to alert
the human triager that there may be an issue with version(s) that needs
to be fixed manually.

Update internal/genericosv.ToReport to populate this field when appropriate,
and add a lint check to ensure that no reports are submitted with this
field populated.

Change-Id: I7af827be44664af2e4b7907570d279203f1a6c3b
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/529276
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>