internal/{report, genericosv}: add optional types to report notes
Add an optional type field to report notes, to allow notes to be annotated
with which tool/process added them (e.g., "lint", "create".)

The notes field can be used by humans or tools to add metadata to a report
that will not be published in the OSV, but is meant as information for
the human triager.

Here it is used by the ToReport function (which converts GHSAs to YAML) to
add notes about errors or warnings that occurred while creating/linting the report,
and need to be fixed by a human.

Change-Id: I9cc37c37dac7171dfbac1af2c147cd491e1e6dbc
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/529837
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>