internal/symbols: compute package paths from file paths
We assume we have the module name and the previous CLs can give us the
root of a module. We should then be able to compute the package import
paths from local paths to package files or directory.

This is part of a series of CLs implementing automatic symbol extraction.

Change-Id: I8bbe83ce8e2f77e51e49c10f622a2e5210dfb8a1
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/534796
Reviewed-by: Tatiana Bradley <tatianabradley@google.com>
Reviewed-by: Jonathan Amsterdam <jba@google.com>
Run-TryBot: Zvonimir Pavlinovic <zpavlinovic@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>