internal/database, deploy: modify validation order for deploy
Instead of checking that the current database index is valid before deploy
(which ensures that an invalid database can't be fixed), check that
the new database to deploy is valid (including with respect to the reports
in the old database) before deploy. If it is valid, deploy it, then check
afterwards that the newly deployed database is valid.

This allows issues caused by, e.g., incomplete deploys, to be fixed
without manual intervention, while still alerting us if there is a
persistent problem.

Change-Id: Ib57901bcada5a4b823d184bef25a64559a8e9e4a
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/541615
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>