deploy, cmd: add check for correct legacy db to deploy
Adds a check that the deployed legacy DB is correct (in addition
to the existing check that the deployed v1 DB is correct).

Also adds more logging to the checkdb/checkdeploy commands so we
know what settings are turned on/off from the deploy logs.

Change-Id: Iffbae9536fc4caef40993987384e256c8d4237ae
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/541676
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>