internal/report: use golden files in lint tests
To make tests easier to update when we make trivial changes (such as the
wording of a lint error), use golden files to hold expected results
for lint tests.

A follow up CL will delete the existing "want" field from the test cases,
but it is left in for now to show that the existing test infra still passes
with this change in place.

Change-Id: Ibeecfef9982bd2bf55d2a699debf931c5d101aa4
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/543160
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>