internal/report: add a lint check that summary contains module/package path
A new lint check ensures that summaries mention at least one of the affected modules or package paths.

As a special case, the check also allows the
summary to mention a prefix of a module/package path, in case the report
affects many paths or a path is very long.

Change-Id: Iff0c9405343b385a8ee471b09df0686ad39360dc
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/543798
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>