internal/worker: make a copy of testdata repo for worker
Instead of directly using testdata folder of the cveschema5 package
in the worker, make a copy of the required test data.

This allows us to transition to a new testing framework in the
cveschema5 package without needing to (immediately) change the worker logic.

For golang/go#49289

Change-Id: Ieab8fb4701ffa30b2dc000a48d86efe524f1cb02
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/545297
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>