internal/report: add basic logic to convert cve5 to report
Add function to convert CVE JSON 5.0 to our internal YAML format.

As a starting point, this function does the same thing as the old function
that converts from CVE JSON 4 to our format. A temporary test checks
that this is the case.

For golang/go#49289

Change-Id: Ie5226537cefaeb9e68b98e5dce9c6b97b29f968a
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/547556
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>