cmd/vulnreport,internal/cveclient: use cve v5 in vulnreport
The vulnreport command now fetches CVE records in JSON 5.0 format
instead of the legacy 4.0 format.

This change also adds a new function, Fetch, which makes an
unauthenticated HTTP request to the CVE5 database to grab a CVE
record.

This function is used in vulnreport instead of the much-slower
cvelistrepo.FetchCVE (which clones the whole cvelistrepo).

Change-Id: Ic255e98d7c1a52301810dc53712fc2ab4a648e70
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/547560
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>