internal/report: add additional test CVEs for cve-to-report
Add additional test CVEs to increase test coverage for the cve-to-report
functions.

Also modify the equivalence test to allow certain fields to be ignored.
This is needed for the Credits field, because v4 doesn't handle it
properly (but there is no point in fixing it because we are deprecating
support for v4).

Change-Id: If103b09ef03e9b676f84c055e3222871eaa4d6d2
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/548055
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>