internal/cvelistrepo: add function List (for CVEs)
Adds function List, which returns a list of CVE ids that have been
added or updated since the given time.

The source of truth for this function is the "delta log"
published in the cvelistV5 repo.

This will be used by the new "stateless" worker to find the most
recent CVEs to triage.

For golang/go#64720

Change-Id: I94369691fa24924b357537aa651da7142449f2a9
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/550855
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>