internal/worker: fix failing test
Test was failing with -pkgsite flag because of some fake data. Replace
the fake module path with a real module path that is known to pkgsite.

(This wasn't picked up by automatic testing because the pkgsite flag
is not set by default.)

Change-Id: I2e16e449605f036dcfe4b14d99f31ab64db8f2e3
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/550856
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>