internal/report: attempt to auto-fix summaries
If there is no module path in a summary,
simply add it to the end, i.e., "<summary> in <module>".

This could result in non-grammatical
phrases so it's meant as a convenience for a human to fix up.

As a last resort, if there is no summary at all, add
the summary '<[alias] | "Vulnerability"> in <module_path>' as a
starting point.

Change-Id: I64810c7c77980654d7973dc605b256e6053c0254
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/576998
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Damien Neil <dneil@google.com>