internal/report: add quick fix for module merging issue
Previously, FixModules would attempt to merge versions even if it
resulted in an invalid version list (e.g., with overlapping ranges).

As a quick fix, bail out from the merge attempt if any of the
results wouldn't make sense, and add a note to the report.

This can be improved in the future by continuing to attempt to merge
modules that can be merged instead of bailing out early.

Change-Id: I5074cd3d04251423da9afa1a52933f0fbeab4b5f
Reviewed-on: https://go-review.googlesource.com/c/vulndb/+/580435
Reviewed-by: Damien Neil <dneil@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>